Address reivewer comments
