+++ sdks/java/io/delta/src/test/java/org/apache/beam/sdk/io/delta/DeltaIOIT.java
@@ -32,14 +32,10 @@ import io.delta.kernel.data.FilteredColumnarBatch;
 import io.delta.kernel.defaults.engine.DefaultEngine;
 import io.delta.kernel.defaults.internal.data.DefaultColumnarBatch;
 import io.delta.kernel.engine.Engine;
-import io.delta.kernel.internal.data.GenericRow;
-import io.delta.kernel.types.BooleanType;
 import io.delta.kernel.types.DataType;
 import io.delta.kernel.types.IntegerType;
 import io.delta.kernel.types.LongType;
-import io.delta.kernel.types.MapType;
 import io.delta.kernel.types.StringType;
-import io.delta.kernel.types.StructField;
 import io.delta.kernel.types.StructType;
 import io.delta.kernel.types.TimestampType;
 import io.delta.kernel.utils.CloseableIterable;
@@ -53,7 +49,6 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
-import javax.annotation.Nullable;
 import org.apache.beam.sdk.managed.Managed;
 import org.apache.beam.sdk.schemas.Schema;
 import org.apache.beam.sdk.testing.PAssert;
@@ -344,10 +339,11 @@ public class DeltaIOIT {
             .addValues(1, "name_1_updated", "update_postimage", 1L, new Instant(123456789000L))
             .build();
 
-    writeCdcCommit(
+    DeltaWriteTestUtils.writeCdcCommit(
         engine,
         repoPath,
         1L,
+        System.currentTimeMillis(),
         deltaSchema,
         null,
         version0FilePath,
@@ -400,217 +396,5 @@ public class DeltaIOIT {
               row.getString(DeltaIO.CHANGE_TYPE_COLUMN),
               row.getInt64(DeltaIO.COMMIT_VERSION_COLUMN)));
     }
-  }
-
-  private static final StructType CDC_ACTION_SCHEMA =
-      new StructType()
-          .add("path", StringType.STRING, false)
-          .add("partitionValues", new MapType(StringType.STRING, StringType.STRING, false), false)
-          .add("size", LongType.LONG, false)
-          .add("dataChange", BooleanType.BOOLEAN, false);
-
-  private static StructType getCustomSingleActionSchema() {
-    StructType originalSchema = io.delta.kernel.internal.actions.SingleAction.FULL_SCHEMA;
-    List<StructField> fields = new ArrayList<>();
-    for (StructField field : originalSchema.fields()) {
-      if (field.getName().equals("cdc")) {
-        fields.add(new StructField("cdc", CDC_ACTION_SCHEMA, true));
-      } else {
-        fields.add(field);
-      }
-    }
-    return new StructType(fields);
-  }
-
-  private static io.delta.kernel.data.Row createSingleAction(
-      StructType customSingleActionSchema, String actionName, io.delta.kernel.data.Row actionRow) {
-    Map<Integer, Object> values = new HashMap<>();
-    values.put(customSingleActionSchema.indexOf(actionName), actionRow);
-    return new GenericRow(customSingleActionSchema, values);
-  }
-
-  private static io.delta.kernel.data.Row createRemoveAction(
-      StructType removeSchema, String path, long deletionTimestamp) {
-    Map<Integer, Object> values = new HashMap<>();
-    values.put(removeSchema.indexOf("path"), path);
-    values.put(removeSchema.indexOf("deletionTimestamp"), deletionTimestamp);
-    values.put(removeSchema.indexOf("dataChange"), true);
-    values.put(removeSchema.indexOf("size"), 100L);
-    return new GenericRow(removeSchema, values);
-  }
-
-  private static io.delta.kernel.data.Row createCdcAction(
-      StructType cdcSchema, String path, long size) {
-    Map<Integer, Object> values = new HashMap<>();
-    values.put(cdcSchema.indexOf("path"), path);
-    values.put(
-        cdcSchema.indexOf("partitionValues"),
-        io.delta.kernel.internal.util.VectorUtils.stringStringMapValue(Collections.emptyMap()));
-    values.put(cdcSchema.indexOf("size"), size);
-    values.put(cdcSchema.indexOf("dataChange"), true);
-    return new GenericRow(cdcSchema, values);
-  }
-
-  private static ColumnVector createColumnVector(
-      List<Row> rows, int fieldIndex, DataType dataType) {
-    return new ColumnVector() {
-      @Override
-      public DataType getDataType() {
-        return dataType;
-      }
-
-      @Override
-      public int getSize() {
-        return rows.size();
-      }
-
-      @Override
-      public void close() {}
-
-      @Override
-      public boolean isNullAt(int rowId) {
-        return rows.get(rowId).getValue(fieldIndex) == null;
-      }
-
-      @Override
-      public boolean getBoolean(int rowId) {
-        return rows.get(rowId).getBoolean(fieldIndex);
-      }
-
-      @Override
-      public int getInt(int rowId) {
-        return rows.get(rowId).getInt32(fieldIndex);
-      }
-
-      @Override
-      public long getLong(int rowId) {
-        if (dataType instanceof TimestampType) {
-          Instant instant = rows.get(rowId).getDateTime(fieldIndex).toInstant();
-          return instant.getMillis() * 1000L;
-        }
-        return rows.get(rowId).getInt64(fieldIndex);
-      }
-
-      @Override
-      public String getString(int rowId) {
-        return rows.get(rowId).getString(fieldIndex);
-      }
-    };
-  }
-
-  /**
-   * Writes a Delta commit containing CDC actions (simulating updates/deletes).
-   *
-   * <p>Note on why this is manual: In a standard Spark or Flink writer, setting the table property
-   * {@code "delta.enableChangeDataFeed" = "true"} automatically instructs the engine to compute and
-   * write the change data files to {@code _change_data/} and append the {@code cdc} actions to the
-   * commit log whenever DML statements (like UPDATE/DELETE) are executed.
-   *
-   * <p>However, we are using the Delta Lake Kernel API which does not contain an SQL execution
-   * engine or a DML parser. Thus, it cannot automatically compute which rows were deleted or
-   * updated. To generate a realistic integration test dataset, we must manually construct these
-   * change records, write them into the GCS {@code _change_data/} directory using the low-level
-   * parquet handler, and manually register them as {@code cdc} actions in the committed
-   * transaction.
-   */
-  private void writeCdcCommit(
-      Engine engine,
-      String tablePath,
-      long expectedVersion,
-      StructType deltaSchema,
-      @Nullable List<Row> addBeamRows,
-      @Nullable String removePath,
-      @Nullable List<Row> cdcBeamRows,
-      StructType cdcWriteSchema)
-      throws Exception {
-
-    Table table = Table.forPath(engine, tablePath);
-    TransactionBuilder txnBuilder =
-        table.createTransactionBuilder(engine, "DeltaIOIT", Operation.WRITE);
-    Transaction txn = txnBuilder.build(engine);
-    io.delta.kernel.data.Row txnState = txn.getTransactionState(engine);
-
-    StructType customSingleActionSchema = getCustomSingleActionSchema();
-    List<io.delta.kernel.data.Row> commitActions = new ArrayList<>();
-
-    if (addBeamRows != null && !addBeamRows.isEmpty()) {
-      ColumnVector[] vectors = new ColumnVector[deltaSchema.fields().size()];
-      for (int i = 0; i < deltaSchema.fields().size(); i++) {
-        StructField field = deltaSchema.fields().get(i);
-        vectors[i] = createColumnVector(addBeamRows, i, field.getDataType());
-      }
-      ColumnarBatch columnarBatch =
-          new DefaultColumnarBatch(addBeamRows.size(), deltaSchema, vectors);
-      FilteredColumnarBatch filteredBatch =
-          new FilteredColumnarBatch(columnarBatch, Optional.empty());
-      CloseableIterator<FilteredColumnarBatch> data =
-          io.delta.kernel.internal.util.Utils.toCloseableIterator(
-              Collections.singletonList(filteredBatch).iterator());
-      CloseableIterator<FilteredColumnarBatch> physicalData =
-          Transaction.transformLogicalData(engine, txnState, data, Collections.emptyMap());
-      DataWriteContext writeContext =
-          Transaction.getWriteContext(engine, txnState, Collections.emptyMap());
-      CloseableIterator<DataFileStatus> dataFiles =
-          engine
-              .getParquetHandler()
-              .writeParquetFiles(
-                  writeContext.getTargetDirectory(),
-                  physicalData,
-                  writeContext.getStatisticsColumns());
-      CloseableIterator<io.delta.kernel.data.Row> addActions =
-          Transaction.generateAppendActions(engine, txnState, dataFiles, writeContext);
-      while (addActions.hasNext()) {
-        commitActions.add(addActions.next());
-      }
-    }
-
-    if (removePath != null) {
-      StructType removeSchema =
-          (StructType)
-              io.delta.kernel.internal.actions.SingleAction.FULL_SCHEMA
-                  .fields()
-                  .get(io.delta.kernel.internal.actions.SingleAction.REMOVE_FILE_ORDINAL)
-                  .getDataType();
-      io.delta.kernel.data.Row removeAction =
-          createRemoveAction(removeSchema, removePath, System.currentTimeMillis());
-      commitActions.add(createSingleAction(customSingleActionSchema, "remove", removeAction));
-    }
-
-    if (cdcBeamRows != null && !cdcBeamRows.isEmpty()) {
-      ColumnVector[] vectors = new ColumnVector[cdcWriteSchema.fields().size()];
-      for (int i = 0; i < cdcWriteSchema.fields().size(); i++) {
-        StructField field = cdcWriteSchema.fields().get(i);
-        vectors[i] = createColumnVector(cdcBeamRows, i, field.getDataType());
-      }
-      ColumnarBatch columnarBatch =
-          new DefaultColumnarBatch(cdcBeamRows.size(), cdcWriteSchema, vectors);
-      FilteredColumnarBatch filteredBatch =
-          new FilteredColumnarBatch(columnarBatch, Optional.empty());
-      CloseableIterator<FilteredColumnarBatch> data =
-          io.delta.kernel.internal.util.Utils.toCloseableIterator(
-              Collections.singletonList(filteredBatch).iterator());
-
-      String cdcDir = tablePath + "/_change_data";
-
-      CloseableIterator<DataFileStatus> cdcFiles =
-          engine.getParquetHandler().writeParquetFiles(cdcDir, data, Collections.emptyList());
-
-      StructType cdcActionSchema = CDC_ACTION_SCHEMA;
-      while (cdcFiles.hasNext()) {
-        DataFileStatus cdcFile = cdcFiles.next();
-        String relativeCdcPath =
-            "_change_data/" + cdcFile.getPath().substring(cdcFile.getPath().lastIndexOf('/') + 1);
-        io.delta.kernel.data.Row cdcAction =
-            createCdcAction(cdcActionSchema, relativeCdcPath, cdcFile.getSize());
-        commitActions.add(createSingleAction(customSingleActionSchema, "cdc", cdcAction));
-      }
-    }
-
-    TransactionCommitResult result =
-        txn.commit(
-            engine,
-            CloseableIterable.inMemoryIterable(
-                io.delta.kernel.internal.util.Utils.toCloseableIterator(commitActions.iterator())));
-    org.junit.Assert.assertEquals(expectedVersion, result.getVersion());
   }
 }
